backend/requirements.txt
aiosmtplib==3.0.2
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
bcrypt==5.0.0
blinker==1.9.0
celery==5.4.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
dnspython==2.8.0
docker==7.1.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
fastapi-mail==1.5.0
flower==2.0.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
python-owasp-zap-v2.4==0.0.22
PyYAML==6.0.3
redis==5.2.1
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rsa==4.9.1
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
#uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
